Move audio system into QtMainApplication directory
