+++ src/QtMainApplication/QtMainApplication.h
@@ -7,10 +7,11 @@
 #include <QToolButton>
 #include <QToolBar>
 #include <QGridLayout>
-#include "Spectrogram/Visualizer/QtSpectrogram.h"
-#include "Spectrogram/Visualizer/QtAudioSystem.h"
 
+#include <Spectrogram/Visualizer/QtSpectrogram.h>
 #include <Spectrogram/Audio/DeviceList.h>
+
+#include "QtAudioSystem.h"
 
 class QtMainApplication : public QMainWindow {
 Q_OBJECT
@@ -20,38 +21,40 @@ public:
 
 public slots:
 
-	void updateSpectrogram(const Audio::Buffer &buffer);
+    void updateSpectrogram(const Audio::Buffer &buffer);
 
-	void updateSources(const DeviceList &deviceList);
+    void updateSources(const DeviceList &deviceList);
 
-	void startButtonClicked();
+    void startButtonClicked();
 
-	void showSaveSuccess(bool success, QString fileName);
+    void showSaveSuccess(bool success, QString fileName);
 
 signals:
-	// These are signals for the plot controls, forwarded
-	// from their respective button press events
-	void startSignal(const Device &device, std::chrono::milliseconds maxLatency, size_t bufferLength);
+
+    // These are signals for the plot controls, forwarded
+    // from their respective button press events
+    void startSignal(const Device &device, std::chrono::milliseconds maxLatency, size_t bufferLength);
+
     void stopSignal();
 
 private:
-	void setupSaveButton();
+    void setupSaveButton();
 
-	QtSpectrogram *spectrogram;
+    QtSpectrogram *spectrogram;
 
-	QToolBar *controls;
-	
-	QToolButton *stopButton; //OPTIONAL, YES CHOSEN
-	QToolButton *startButton; //OPTIONAL, YES CHOSEN
+    QToolBar *controls;
+
+    QToolButton *stopButton;
+    QToolButton *startButton;
 
     QComboBox *audioSelectBox;
 
-	QPushButton *saveButton;
-	QStateMachine saveMachine;
-	QState *off;
-	QState *on;
+    QPushButton *saveButton;
+    QStateMachine saveMachine;
+    QState *off;
+    QState *on;
 
-	const DeviceList *devices = nullptr;
+    const DeviceList *devices = nullptr;
 
 };
